agent: handle session token with the client (API change)
To setup the session token the client must call 'authenticate' to login against
the agent. If the authentication is successfull the agent return a Cookie with
the token information. The client needs to store that cookie and supply the
token in the headers of any following request.

Once the client is authenticated the agent loads the modules depending on the
authentication mode (local / api). The load method is no more exposed through
the XML-RPC API.