www: don't redirect the page on connection lost
When we lost the connection with the agent and we
get it back we check if the agent was restarted or
not using mss.agent.START_TIME. If the agent was
restarded we redirect the application to /mss/ because
we loose our Token with the agent so we need to
authenticate. Otherwise we don't need to redirect.